project-2/backend/requirements.txt
langgraph>=0.2.0
openai>=1.0.0
pillow>=10.0.0
ffmpeg-python>=0.2.0
moviepy>=1.0.3
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0
flask==3.1.1
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.0
uuid>=1.30
requests>=2.31.0
sseclient-py>=1.8.0
pyyaml>=6.0
# Test dependencies
pytest>=7.4.0
pytest-cov>=4.1.0